+++ R/get_activity_info.R
@@ -1,9 +1,14 @@
-#' @title eLTER get_activity_info function
-#' @description This function obtains the info of activity (actually
-#'  title) provided in DEIMS-SDR.
-#' @param activityid A character. The DEIMS activity ID from
-#' DEIMS-SDR website. More information about DEIMS iD in this
-#' \href{https://deims.org/docs/deimsid.html}{page}.
+#' Obtain the information about of an eLTER activity.
+#' @description This function obtains the information about of an eLTER
+#' activity (e.g.
+#' \url{https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845})
+#' provided in \href{https://deims.org/}{DEIMS-SDR catalogue}.
+#' @param activityid A `character`. It is the DEIMS ID of activity make from
+#' DEIMS-SDR website. DEIMS ID information
+#' \href{https://deims.org/docs/deimsid.html}{here}.
+#' The DEIMS.iD of activity is the URL for the activity page.
+#' @param show_map A `boolean`. If TRUE a Leaflet map with occurrences
+#' is shown. Default FALSE.
 #' @return The output of the function is a `tibble` with main features of
 #' the activities in a site, and a `leaflet` map plot.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
@@ -16,16 +21,31 @@
 #' @examples
 #' activities <- get_activity_info(
 #'   activityid =
-#'   "https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845"
+#'   "https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845",
+#'   show_map = TRUE
 #' )
 #' activities
 #'
+#' @section The function output:
+#' \figure{get_activity_info_fig.png}{Map of "Study of non-indigenous (alien)
+#' species in the Mar Piccolo of Taranto" activity}
+#'
 ### function get_activity_info
-get_activity_info <- function(activityid) {
-  q <- "{
+get_activity_info <- function(activityid, show_map = FALSE) {
+  q <- '{
         title: .title,
-        boundaries: .attributes.geographic.boundaries
-        }"
+        abstract: .attributes.general.abstract,
+        keywords: .attributes.general.keywords,
+        uri: "\\(.id.prefix)\\(.id.suffix)",
+        type: .type,
+        created: .created,
+        changed: .changed,
+        relatedSite: .attributes.general.relatedSite,
+        contacts: .attributes.contact,
+        boundaries: .attributes.geographic.boundaries,
+        observationParameters: .attributes.observations.parameters,
+        relatedResources: .attributes.relatedResources
+        }'
   jj <- get_id(activityid, "activities")
   if (is.na(attr(jj, "status"))) {
     invisible(
@@ -47,10 +67,14 @@ get_activity_info <- function(activityid) {
         geoActivity_valid <- sf::st_is_valid(geoActivity)
         if (any(geoActivity_valid)) {
           map <- leaflet::leaflet(geoActivity) %>%
-            leaflet::addTiles() %>%
-            leaflet::addPolygons()
-          print(map)
-          geoActivity
+            leaflet::addTiles()
+            if (sf::st_geometry_type(geoActivity) == "POINT") {
+              map <- map %>%
+                leaflet::addMarkers()
+            } else if (sf::st_geometry_type(geoActivity) == "POLYGON") {
+              map <- map %>%
+                leaflet::addPolygons()
+            }
         } else {
           map <- leaflet::leaflet() %>%
             leaflet::addTiles()
@@ -58,8 +82,6 @@ get_activity_info <- function(activityid) {
 activity, provided in DEIMS-SDR, has an invalid geometry.
 Please check the content and refers this error to DEIMS-SDR
 contact person of the activity, citing the Activity.iD.\n----\n")
-          print(map)
-          geoActivity
         }
       }
     } else {
@@ -70,6 +92,10 @@ contact person of the activity, citing the Activity.iD.\n----\n")
     stop("\n----\nPage Not Found. The requested page could not be found. Please
 check again the Activity.iD\n----\n")
   }
-  print(map)
-  geoActivity
+  if (show_map == TRUE) {
+    print(map)
+    geoActivity
+  } else {
+    geoActivity
+  }
 }