+++ R/get_network_sites.R
@@ -1,28 +1,39 @@
-#' @title eLTER get_network_sites function
-#' @description This function retrieves a list of sites in the provided network
-#' including title, date late updated, URI, and coordinates
+#' Retrieve a list of sites in an eLTER Network.
+#' @description This function return a spatial point vector object including
+#' title, date late updated, URI, and coordinates, stored in
+#' \href{https://deims.org/}{DEIMS-SDR catalogue}, of all the eLTER sites
+#' belonging to an eLTER Network (e.g.
+#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
+#' Italy network}).
 #' @param networkDEIMSID A `character`. The DEIMS ID of the network from
-#' DEIMS-SDR website. More information about DEIMS network ID from these pages:
-#' \href{https://deims.org/docs/deimsid.html}{page}, and
-#' \href{https://deims.org/search?f[0]=result_type:network}{page} the
-#' complete list of ILTER networks.
-#'
+#' DEIMS-SDR website. DEIMS ID information
+#' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of
+#' networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
+#' The DEIMS ID of network is the URL for the network page.
 #' @return The output of the function is a point vector of `sf` class
 #' (package sf) of the network's sites.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom jsonlite fromJSON
-#' @importFrom sf st_as_sf st_is_valid
+#' @importFrom sf st_as_sf st_is_valid st_cast
 #' @importFrom dplyr select as_tibble
 #' @importFrom leaflet leaflet addTiles addMarkers
 #' @importFrom httr RETRY content
 #' @export
 #' @examples
 #' \dontrun{
+#' # The sites of LTER-Italy network
 #' listSites <- get_network_sites(
 #'   networkDEIMSID =
 #'   "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
 #' )
-#' listSites[1:10, ]
+#' listSites
+#'
+#' # The sites of LTER Europe network
+#' euSites <- get_network_sites(
+#'   networkDEIMSID =
+#'   "https://deims.org/networks/4742ffca-65ac-4aae-815f-83738500a1fc"
+#' )
+#' euSites
 #' }
 #'
 ### function get_network_sites
@@ -38,7 +49,8 @@ get_network_sites <- function(networkDEIMSID) {
   if (length(lterNetworkSitesCoords) != 0) {
     lterSitesNetworkPointDEIMS <- sf::st_as_sf(
       lterNetworkSitesCoords,
-      wkt = "coordinates"
+      wkt = "coordinates",
+      crs = 4326
     )
     lterSitesNetworkPointDEIMS$uri <- paste0(
       lterSitesNetworkPointDEIMS$id$prefix,
@@ -51,11 +63,54 @@ get_network_sites <- function(networkDEIMSID) {
     lSNPD_valid <- sf::st_is_valid(
       lterSitesNetworkPointDEIMS
     )
+  
+    # checking MULTIPOINT geometry
+    lSNPD_type <- sf::st_geometry_type(
+      x = lterSitesNetworkPointDEIMS,
+      by_geometry = TRUE
+    )
+    suppressWarnings(
+    if (isTRUE(any(match("MULTIPOINT", lSNPD_type)) != "POINT")) {
+      lSNPD_temp <- NULL
+      for (i in match("MULTIPOINT", lSNPD_type)) {
+        lSNPD_splitted <- sf::st_cast(
+          x = lterSitesNetworkPointDEIMS[c(i), ],
+          to = "POINT",
+          do_split = TRUE
+        )
+        lSNPD_temp <- rbind(
+          lSNPD_temp,
+          lSNPD_splitted
+        )
+        lterSitesNetworkPointDEIMS <- lterSitesNetworkPointDEIMS[c(-i), ]
+      }
+      lterSitesNetworkPointDEIMS <- rbind(
+        lterSitesNetworkPointDEIMS,
+        lSNPD_temp
+      )
+    })
+    # end checking
+  
     if (any(lSNPD_valid)) {
       map <- leaflet::leaflet(lterSitesNetworkPointDEIMS) %>%
         leaflet::addTiles() %>%
-        leaflet::addMarkers()
+        leaflet::addMarkers(
+          clusterOptions = leaflet::markerClusterOptions(),
+          popup = paste0(
+            "<b>Site name: </b>"
+            , lterSitesNetworkPointDEIMS$title
+            , "<br>"
+            , "<a href='"
+            , lterSitesNetworkPointDEIMS$uri
+            , "' target='_blank'>"
+            , "Click Here to View site landing page</a>"
+          )
+        )
       print(map)
+      message("\n----\nThe number of the sites on the map can be more than
+presents in the network, because some theme are represented in DEIMS-SDR
+with multiple points (e.g.
+https://deims.org/18998d9a-7ff5-4e9d-a971-9694e0a4914d).\n----\n")
       return(lterSitesNetworkPointDEIMS)
     } else {
       message("\n----\nThe maps cannot be created because the coordinates,