+++ R/get_site_ODS.R
@@ -1,13 +1,13 @@
-#' @title eLTER get_site_ODS function
-#' @description This function acquires various raster layers from
-#' ODS Europe:  https://maps.opendatascience.eu/
-#' and crops to an eLTER site boundary, which is obtained
-#' from the DEIMS-SDR sites API.
-#'
-#' @param deimsid  a `character`. The DEIMS ID of the site from
-#' DEIMS-SDR website. More information about DEIMS ID from:
-#' \href{https://deims.org/docs/deimsid.html}{page}.
-#' @param dataset a `character`. The requested dataset. One of:
+#' Acquire various raster layers from
+#' \href{https://maps.opendatascience.eu/}{ODS Europe}
+#' and crops to an eLTER site boundary.
+#' @description Download and return a SpatRaster object containing the requested
+#' dataset from \href{https://maps.opendatascience.eu/}{ODS},
+#' cropped to an eLTER site boundary, which is obtained from the DEIMS-SDR API.
+#' @param deimsid  A `character`. The DEIMS ID of the site from
+#' DEIMS-SDR website. DEIMS ID information
+#' \href{https://deims.org/docs/deimsid.html}{here}.
+#' @param dataset A `character`. The requested dataset. One of:
 #' "landcover", "clc2018", "osm_buildings", "natura2000",
 #' "ndvi_spring", "ndvi_summer", "ndvi_autumn", "ndvi_winter".
 #' Default is "landcover".
@@ -62,6 +62,9 @@
 #' siteNDVI
 #' terra::plot(siteNDVI)
 #' }
+#'
+#' @section The function output:
+#' \figure{get_site_ods_fig.png}{NDVI for Eisenwurzen}
 #'
 ### function get_site_ODS
 get_site_ODS <- function(deimsid, dataset = "landcover") {
@@ -128,5 +131,9 @@ get_site_ODS <- function(deimsid, dataset = "landcover") {
   boundary <- sf::st_transform(boundary, terra::crs(ds))
   boundary <- terra::vect(boundary)
   ds_site <- terra::mask(terra::crop(ds, boundary), boundary)
+  # If this is NDVI, rescale back to (-1.0,1.0) range
+  if (length(grep(pattern = "ndvi", x = dataset, fixed = TRUE)) > 0) {
+    ds_site <- (ds_site - 100) / 100.0
+  }
   return(ds_site)
 }