+++ R/get_site_general.R
@@ -2,8 +2,8 @@
 #' @description This function obtains general information
 #' about an eLTER site through the DEIMS-SDR sites API.
 #' @param deimsid A `character`. The DEIMS ID of the site from
-#' DEIMS-SDR website. More information about DEIMS ID from:
-#' \href{https://deims.org/docs/deimsid.html}{page}.
+#' DEIMS-SDR website. DEIMS ID information
+#' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a `tibble` with main features of the
 #' site and the general information, such as: abstract, keywords, purpose,
 #' status, yearEstablished, yearClosed, hierarchy, siteName, short name, site
@@ -38,7 +38,7 @@ get_site_general <- function(deimsid) {
     )
   } else {
     message("\n----\nThe requested page could not be found.
-Please check again the DEIMS.iD\n----\n")
+Please check again the DEIMS ID\n----\n")
     general <- NULL
   }
   general