+++ R/onAttach.R
@@ -6,7 +6,7 @@
     developers (https://github.com/oggioniale/ReLTER).\n\n",
     "If you use this package, please cite as:\n\n",
     "Alessandro Oggioni, Micha Silver, Luigi Ranghetti & Paolo Tagliolato.
-    (2021) oggioniale/ReLTER: ReLTER v1.0.0 (1.0.0). Zenodo.
+    (2022) oggioniale/ReLTER: ReLTER v1.1.0 (1.1.0). Zenodo.
     https://doi.org/10.5281/zenodo.5576813\n\n",
     "Type 'citation(package = 'ReLTER')' on how to cite R packages in
     publications."